requirements.txt
flask
python-dotenv
groq
google-genai
